Fix Marlon shop when calling him
