+++ StardewArchipelago/Locations/Patcher/ModLocationPatcher.cs
@@ -79,7 +79,7 @@ namespace StardewArchipelago.Locations.Patcher
             {
                 _harmony.Patch(
                     original: AccessTools.Method(typeof(ShopMenu), nameof(ShopMenu.update)),
-                    postfix: new HarmonyMethod(typeof(MagicModInjections), nameof(MagicModInjections.Update_ReplaceMarlonRecipes_Postfix))
+                    postfix: new HarmonyMethod(typeof(MagicModInjections), nameof(MagicModInjections.Update_ReplaceMarlonShopChecks_Postfix))
             );
             }
             /*_harmony.Patch(
